Extract Msf::Module::FullName
MSP-11126
