+++ lib/msf/core/module.rb
@@ -18,6 +18,7 @@ class Module
   autoload :Compatibility, 'msf/core/module/compatibility'
   autoload :DataStore, 'msf/core/module/data_store'
   autoload :Deprecated, 'msf/core/module/deprecated'
+  autoload :FullName, 'msf/core/module/full_name'
   autoload :HasActions, 'msf/core/module/has_actions'
   autoload :ModuleInfo, 'msf/core/module/module_info'
   autoload :ModuleStore, 'msf/core/module/module_store'
@@ -34,6 +35,7 @@ class Module
   include Msf::Module::Arch
   include Msf::Module::Compatibility
   include Msf::Module::DataStore
+  include Msf::Module::FullName
   include Msf::Module::ModuleInfo
   include Msf::Module::ModuleStore
   include Msf::Module::Options
@@ -46,20 +48,6 @@ class Module
 
   class << self
     include Framework::Offspring
-
-    def fullname
-      type + '/' + refname
-    end
-
-    def shortname
-      refname.split('/').last
-    end
-
-    #
-    # The module's name that is assigned it it by the framework
-    # or derived from the path that the module is loaded from.
-    #
-    attr_accessor :refname
 
     #
     # This attribute holds the non-duplicated copy of the module
@@ -157,38 +145,6 @@ class Module
     obj.user_output  = self.user_output
     obj.module_store = self.module_store.clone
     obj
-  end
-
-  #
-  # Returns the module's framework full reference name.  This is the
-  # short name that end-users work with (refname) plus the type
-  # of module prepended.  Ex:
-  #
-  # payloads/windows/shell/reverse_tcp
-  #
-  def fullname
-    self.class.fullname
-  end
-
-  #
-  # Returns the module's framework reference name.  This is the
-  # short name that end-users work with.  Ex:
-  #
-  # windows/shell/reverse_tcp
-  #
-  def refname
-    self.class.refname
-  end
-
-  #
-  # Returns the module's framework short name.  This is a
-  # possibly conflicting name used for things like console
-  # prompts.
-  #
-  # reverse_tcp
-  #
-  def shortname
-    self.class.shortname
   end
 
   #
